+++ src/format_level_lowercase.cpp
@@ -26,32 +26,33 @@ namespace uuid {
 
 namespace log {
 
-static const char *pstr_level_lowercase_off __attribute__((__aligned__(sizeof(int)))) PROGMEM = "off";
-static const char *pstr_level_lowercase_emerg __attribute__((__aligned__(sizeof(int)))) PROGMEM = "emerg";
-static const char *pstr_level_lowercase_crit __attribute__((__aligned__(sizeof(int)))) PROGMEM = "crit";
-static const char *pstr_level_lowercase_alert __attribute__((__aligned__(sizeof(int)))) PROGMEM = "alert";
-static const char *pstr_level_lowercase_err __attribute__((__aligned__(sizeof(int)))) PROGMEM = "err";
-static const char *pstr_level_lowercase_warning __attribute__((__aligned__(sizeof(int)))) PROGMEM = "warning";
-static const char *pstr_level_lowercase_notice __attribute__((__aligned__(sizeof(int)))) PROGMEM = "notice";
-static const char *pstr_level_lowercase_info __attribute__((__aligned__(sizeof(int)))) PROGMEM = "info";
-static const char *pstr_level_lowercase_debug __attribute__((__aligned__(sizeof(int)))) PROGMEM = "debug";
-static const char *pstr_level_lowercase_trace __attribute__((__aligned__(sizeof(int)))) PROGMEM = "trace";
-static const char *pstr_level_lowercase_all __attribute__((__aligned__(sizeof(int)))) PROGMEM = "all";
+static constexpr const char *pstr_level_lowercase_off __attribute__((__aligned__(sizeof(int)))) PROGMEM = "off";
+static constexpr const char *pstr_level_lowercase_emerg __attribute__((__aligned__(sizeof(int)))) PROGMEM = "emerg";
+static constexpr const char *pstr_level_lowercase_crit __attribute__((__aligned__(sizeof(int)))) PROGMEM = "crit";
+static constexpr const char *pstr_level_lowercase_alert __attribute__((__aligned__(sizeof(int)))) PROGMEM = "alert";
+static constexpr const char *pstr_level_lowercase_err __attribute__((__aligned__(sizeof(int)))) PROGMEM = "err";
+static constexpr const char *pstr_level_lowercase_warning __attribute__((__aligned__(sizeof(int)))) PROGMEM = "warning";
+static constexpr const char *pstr_level_lowercase_notice __attribute__((__aligned__(sizeof(int)))) PROGMEM = "notice";
+static constexpr const char *pstr_level_lowercase_info __attribute__((__aligned__(sizeof(int)))) PROGMEM = "info";
+static constexpr const char *pstr_level_lowercase_debug __attribute__((__aligned__(sizeof(int)))) PROGMEM = "debug";
+static constexpr const char *pstr_level_lowercase_trace __attribute__((__aligned__(sizeof(int)))) PROGMEM = "trace";
+static constexpr const char *pstr_level_lowercase_all __attribute__((__aligned__(sizeof(int)))) PROGMEM = "all";
+
+static constexpr const __FlashStringHelper *log_level_lowercase[(int)Level::ALL - (int)Level::OFF + 1] __attribute__((__aligned__(sizeof(int)))) PROGMEM = {
+	reinterpret_cast<const __FlashStringHelper *>(pstr_level_lowercase_off),
+	reinterpret_cast<const __FlashStringHelper *>(pstr_level_lowercase_emerg),
+	reinterpret_cast<const __FlashStringHelper *>(pstr_level_lowercase_crit),
+	reinterpret_cast<const __FlashStringHelper *>(pstr_level_lowercase_alert),
+	reinterpret_cast<const __FlashStringHelper *>(pstr_level_lowercase_err),
+	reinterpret_cast<const __FlashStringHelper *>(pstr_level_lowercase_warning),
+	reinterpret_cast<const __FlashStringHelper *>(pstr_level_lowercase_notice),
+	reinterpret_cast<const __FlashStringHelper *>(pstr_level_lowercase_info),
+	reinterpret_cast<const __FlashStringHelper *>(pstr_level_lowercase_debug),
+	reinterpret_cast<const __FlashStringHelper *>(pstr_level_lowercase_trace),
+	reinterpret_cast<const __FlashStringHelper *>(pstr_level_lowercase_all)
+};
 
 const __FlashStringHelper *format_level_lowercase(Level level) {
-	const __FlashStringHelper *log_level_lowercase[(int)Level::ALL - (int)Level::OFF + 1] = {
-			reinterpret_cast<const __FlashStringHelper *>(pstr_level_lowercase_off),
-			reinterpret_cast<const __FlashStringHelper *>(pstr_level_lowercase_emerg),
-			reinterpret_cast<const __FlashStringHelper *>(pstr_level_lowercase_crit),
-			reinterpret_cast<const __FlashStringHelper *>(pstr_level_lowercase_alert),
-			reinterpret_cast<const __FlashStringHelper *>(pstr_level_lowercase_err),
-			reinterpret_cast<const __FlashStringHelper *>(pstr_level_lowercase_warning),
-			reinterpret_cast<const __FlashStringHelper *>(pstr_level_lowercase_notice),
-			reinterpret_cast<const __FlashStringHelper *>(pstr_level_lowercase_info),
-			reinterpret_cast<const __FlashStringHelper *>(pstr_level_lowercase_debug),
-			reinterpret_cast<const __FlashStringHelper *>(pstr_level_lowercase_trace),
-			reinterpret_cast<const __FlashStringHelper *>(pstr_level_lowercase_all)
-	};
 	return log_level_lowercase[(int)level + 1];
 }